+++ Project1/TuitionUp.h
@@ -36,11 +36,11 @@ public:
 
     /// Get the width of the item
     /// \returns item width
-    virtual int GetWidth() override { return GetImage()->GetWidth(); };
+    virtual double GetWidth() override { return GetImage()->GetWidth(); };
 
     /// Get the height of the item
     /// \returns item height
-    virtual int GetHeight() override { return GetImage()->GetHeight(); };
+    virtual double GetHeight() override { return GetImage()->GetHeight(); };
 
 private:
